+++ src/index.core.ts
@@ -1,4 +1,0 @@
-import QuillCursors from './quill-cursors/quill-cursors';
-import Cursor from './quill-cursors/cursor';
-
-export {QuillCursors as default, Cursor};